backend-riego-final/requirements.txt
Django>=5.0,<6.0
mysqlclient>=2.2.0
djangorestframework>=3.15.0
djangorestframework-simplejwt>=5.5.0
gunicorn>=23.0.0